Standardize DTO naming and consolidate workflow request DTOs
- Rename GenerateWorkflowRequest to GenerateWorkflowRequestDto for consistency
- Move workflow-related DTOs to Application.Contracts/WorkflowOrchestration/WorkflowRequestDto.cs
- Consolidate GenerateWorkflowRequestDto, TextCompletionRequestDto, and TextCompletionResponseDto in one file
- Add consistent validation attributes (Required, MinLength) to both request DTOs
- Update ITextCompletionService and TextCompletionService to use new DTO namespace
- Update WorkflowController to use new DTO types
- Remove old TextCompletionServiceDto.cs file

All DTOs now follow consistent naming pattern (*Dto suffix) and are organized in logical modules.
