+++ station/src/Aevatar.Application.Contracts/Service/ITextCompletionService.cs
@@ -1,4 +1,5 @@
 using System.Threading.Tasks;
+using Aevatar.Application.Contracts.WorkflowOrchestration;
 
 namespace Aevatar.Service;
