+++ station/src/Aevatar.Application/Service/TextCompletionService.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Aevatar.Application.Contracts.WorkflowOrchestration;
 using Aevatar.Application.Grains.Agents.AI;
 using Aevatar.Core.Abstractions;
 using Aevatar.GAgents.AIGAgent.Dtos;
